add data: trends, tweets, followers
