+++ src/Main/SidebarHeading.js
