+++ src/Main/Trends/index.js
@@ -41,39 +41,7 @@ const List = styled.ul`
   margin: 0;
 `;
 
-const trends = [
-  {
-    topic: '#BringYourDogToWorkDay',
-    to: '/linktrend',
-  },
-  {
-    topic: '#FridayFeeling"',
-    to: '/linktrend',
-    count: 12100,
-  },
-  {
-    topic: '#BrexitAnniversary',
-    to: '/linktrend',
-    description: 'It’s one year since the UK voted to leave the European Union',
-  },
-  {
-    topic: 'HMS Queen Elizabeth',
-    to: '/linktrend',
-    count: 1036,
-  },
-  {
-    topic: 'Joe Budden',
-    to: '/linktrend',
-    count: 1036,
-  },
-  {
-    topic: 'Trident',
-    to: '/linktrend',
-    count: 12100,
-  },
-];
-
-export default () => (
+export default ({ data }) => (
   <Trends>
     <Heading>
       <Title>
@@ -85,7 +53,7 @@ Change
       </Button>
     </Heading>
     <List>
-      {trends.map(trend => (
+      {data.map(trend => (
         <Trend
           topic={trend.topic}
           to={trend.to}
