--- src/ProfilePage/Header/index.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import Navigation from './Navigation';
-
-export default () => (
-  <div>
-    <Navigation />
-  </div>
-);
