+++ src/data/followers.js
@@ -1,0 +1,26 @@
+export default [
+  {
+    to: '/follower1',
+    avatar: `${process.env.PUBLIC_URL}/img/avatar-1.jpg`,
+  },
+  {
+    to: '/follower2',
+    avatar: `${process.env.PUBLIC_URL}/img/avatar-2.jpg`,
+  },
+  {
+    to: '/follower3',
+    avatar: `${process.env.PUBLIC_URL}/img/avatar-3.jpg`,
+  },
+  {
+    to: '/follower4',
+    avatar: `${process.env.PUBLIC_URL}/img/avatar-4.jpg`,
+  },
+  {
+    to: '/follower5',
+    avatar: `${process.env.PUBLIC_URL}/img/avatar-5.jpg`,
+  },
+  {
+    to: '/follower6',
+    avatar: `${process.env.PUBLIC_URL}/img/avatar-6.jpg`,
+  },
+];
