+++ src/data/trends.js
@@ -1,0 +1,31 @@
+export default [
+  {
+    topic: '#BringYourDogToWorkDay',
+    to: '/linktrend',
+  },
+  {
+    topic: '#FridayFeeling"',
+    to: '/linktrend',
+    count: 12100,
+  },
+  {
+    topic: '#BrexitAnniversary',
+    to: '/linktrend',
+    description: 'It’s one year since the UK voted to leave the European Union',
+  },
+  {
+    topic: 'HMS Queen Elizabeth',
+    to: '/linktrend',
+    count: 1036,
+  },
+  {
+    topic: 'Joe Budden',
+    to: '/linktrend',
+    count: 1036,
+  },
+  {
+    topic: 'Trident',
+    to: '/linktrend',
+    count: 12100,
+  },
+];
