+++ src/data/tweets.js
@@ -1,0 +1,38 @@
+export default [
+  {
+    pinned: true,
+    nick: 'Every Interaction',
+    user: 'EveryInteract',
+    avatar: `${process.env.PUBLIC_URL}/img/avatar.jpg`,
+    time: '2018-03-12T00:00:00+03:00',
+    text:
+      'We’ve made some more resources for all you wonderful #design folk everyinteraction.com/resources/ #webdesign #UI',
+    img: `${process.env.PUBLIC_URL}/img/post.jpg`,
+    replies: 0,
+    retweets: 17,
+    likes: 47,
+  },
+  {
+    nick: 'Every Interaction',
+    user: 'EveryInteract',
+    avatar: `${process.env.PUBLIC_URL}/img/avatar.jpg`,
+    time: '2018-06-21T00:00:00+03:00',
+    text:
+      'Our new website concept; Taking you from… @ Every Interaction instagram.com/p/BNFGrfhBP3M/',
+    replies: 1,
+    retweets: 4,
+    loves: 2,
+  },
+  {
+    nick: 'Every Interaction',
+    user: 'EveryInteract',
+    avatar: `${process.env.PUBLIC_URL}/img/avatar.jpg`,
+    time: '2018-11-18T00:00:00+03:00',
+    text:
+      'Variable web fonts are coming, and will open a world of opportunities for weight use online',
+    img: `${process.env.PUBLIC_URL}/img/post3.jpg`,
+    replies: 0,
+    retweets: 0,
+    loves: 0,
+  },
+];
